+++ src/components/homepage/Movies-App.js
@@ -138,6 +138,80 @@ const movies3 =[
 }
 ]
 
+const cinemas = [
+  {
+    name: "Gopalan Cinemas",
+    imgURL:
+      "https://www.boxofficepro.com/wp-content/uploads/2019/03/Cinepolis-Forum-Shantiniketan-Bengaluru.jpg",
+    timings: [
+      "9:30 am",
+      "12:15 pm",
+      "2:30 pm",
+      "4:00 pm",
+      "7:20 pm",
+      "9:45 pm"
+    ],
+    format: ["2D", "3D", "IMAX 3D"]
+  },
+  {
+    name: "PVR Vega City",
+    imgURL:
+      "https://www.boxofficepro.com/wp-content/uploads/2019/03/Cinepolis-Forum-Shantiniketan-Bengaluru.jpg",
+    timings: [
+      "9:30 am",
+      "12:15 pm",
+      "2:30 pm",
+      "4:00 pm",
+      "7:20 pm",
+      "9:45 pm"
+    ],
+    format: ["2D", "3D", "IMAX 3D"]
+  },
+
+  {
+    name: "Cinépolis Nexus Shantiniketan",
+    imgURL:
+      "https://www.boxofficepro.com/wp-content/uploads/2019/03/Cinepolis-Forum-Shantiniketan-Bengaluru.jpg",
+      timings: [
+        "9:30 am",
+        "12:15 pm",
+        "2:30 pm",
+        "4:00 pm",
+        "7:20 pm",
+        "9:45 pm"
+      ],
+    format: ["2D", "3D", "IMAX 3D"]
+  },
+  {
+    name: "Inox JP Nagar",
+    imgURL:
+      "https://www.boxofficepro.com/wp-content/uploads/2019/03/Cinepolis-Forum-Shantiniketan-Bengaluru.jpg",
+      timings: [
+        "9:30 am",
+        "12:15 pm",
+        "2:30 pm",
+        "4:00 pm",
+        "7:20 pm",
+        "9:45 pm"
+      ],
+    format: ["2D", "3D","2D", "3D"]
+  },
+  {
+    name: "PVR Koramangala",
+    imgURL:
+      "https://www.boxofficepro.com/wp-content/uploads/2019/03/Cinepolis-Forum-Shantiniketan-Bengaluru.jpg",
+    timings: [
+      "9:30 am",
+      "12:15 pm",
+      "2:30 pm",
+      "4:00 pm",
+      "7:20 pm",
+      "9:45 pm"
+    ],
+    format: ["2D", "3D", "IMAX 3D"]
+  }
+];
+
 function makeMovie(mv,index){
   return(
     <Movie
@@ -171,4 +245,7 @@ function MvApp() {
   );
 }
 
-export default MvApp;
+export {
+  MvApp,
+  cinemas,
+};
